Implemented job activation per lifetime scope
Hangfire does not support per request containers. This means that all
dependencies will be shared between jobs, which is far from ideal,
especially when you have IDbContext. This is on the road map of Hangfire
2.0.
We need a work around this. This is why we add our own
AutofacJobActivator and a JobFilter to manage a per job lifetime scope.